+++ src/app/task.routes.ts
@@ -4,26 +4,24 @@ export const taskRoutes: Routes = [
   {
     path: '',
     pathMatch: 'full',
-    loadComponent: () => import('./tasks/pages/tasks/tasks.component')
+    loadComponent: () => import('./tasks/pages/tasks/tasks.component'),
   },
   {
     path: 'new',
-    loadComponent: () => import('./tasks/pages/new-task/new-task.component')
+    loadComponent: () => import('./tasks/pages/new-task/new-task.component'),
   },
   {
     path: ':id/edit',
-    loadComponent: () => import('./tasks/pages/update-task/update-task.component')
+    loadComponent: () =>
+      import('./tasks/pages/update-task/update-task.component'),
   },
   {
     path: ':id',
-    loadComponent: () => import('./tasks/pages/task-details/task-details.component')
-  },
-  {
-    path: 'state/:state',
-    loadComponent: () => import('./tasks/pages/tasks-by-state/tasks-by-state.component')
+    loadComponent: () =>
+      import('./tasks/pages/task-details/task-details.component'),
   },
   {
     path: '**',
-    redirectTo: ''
-  }
+    redirectTo: '',
+  },
 ];