+++ src/app/tasks/pages/task-details/task-details.component.ts
@@ -1,9 +1,19 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
+import { headerTitle$ } from '../../components/header/header.component';
 
 @Component({
   selector: 'app-task-details',
-  imports: [],
+  imports: [TranslateModule],
   templateUrl: './task-details.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export default class TaskDetailsComponent {}
+export default class TaskDetailsComponent {
+  translate = inject(TranslateService);
+
+  ngOnInit(): void {
+    // TODO: Try to avoid this life cycle
+    const headerTitle = this.translate.instant('taskDetails.title');
+    headerTitle$.next(headerTitle);
+  }
+}