+++ src/app/tasks/pages/tasks-by-state/tasks-by-state.component.ts
@@ -1,9 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-
-@Component({
-  selector: 'app-tasks-by-state',
-  imports: [],
-  templateUrl: './tasks-by-state.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export default class TasksByStateComponent {}